feat: layerTree withSearch
