+++ apps/cim-map/src/pages/Resource/index.js
@@ -1,11 +1,13 @@
 import { useViewStore } from '@chu/store';
 import { Panel } from '@chu/ui';
-import { LayerList, LayerTree, Legend } from '@chu/widgets';
+import { LayerList, LayerTree, withSearch, Legend } from '@chu/widgets';
 import { Flex } from 'antd';
 import { useEffect, useState } from 'react';
 import styles from './index.less';
 import { getLayerTree } from './service.js';
 import getLayerInfo from '@/utils/getLayerInfo';
+
+const Tree = withSearch(LayerTree);
 
 const ResourcePage = () => {
   const view = useViewStore((state) => state.view);
@@ -18,7 +20,7 @@ const ResourcePage = () => {
       <div className={styles.left}>
         <Flex gap="large" vertical>
           <Panel title="目录树">
-            <LayerTree view={view} treeData={treeData} getLayerInfo={getLayerInfo} />
+            <Tree treeData={treeData} getLayerInfo={getLayerInfo} />
           </Panel>
         </Flex>
       </div>
