+++ packages/widgets/src/LayerTree/index.jsx
@@ -1,20 +1,3 @@
-import { layerControl } from '@chu/middleware';
-import { layerTreeStoreCreator, withMiddlewares } from '@chu/store';
-import { Tree } from 'antd';
-import { useMemo } from 'react';
+export { default as withSearch } from './withSearch';
 
-const LayerTree = ({ treeData, getLayerInfo }) => {
-  const useLayerTreeStore = useMemo(
-    () => withMiddlewares(layerTreeStoreCreator, [layerControl(getLayerInfo(treeData))]),
-    [treeData, getLayerInfo],
-  );
-  const { checkedKeys, setCheckedKeys } = useLayerTreeStore();
-
-  const onCheck = (checkedKeysValue) => {
-    setCheckedKeys(checkedKeysValue);
-  };
-
-  return <Tree checkable onCheck={onCheck} checkedKeys={checkedKeys} treeData={treeData} />;
-};
-
-export default LayerTree;
+export { default as LayerTree } from './core';
