+++ packages/widgets/src/index.js
@@ -1,6 +1,6 @@
 export { default as Daylight } from './Daylight';
 export { default as LayerList } from './LayerList';
-export { default as LayerTree } from './LayerTree/search';
+export { LayerTree, withSearch } from './LayerTree';
 export { default as Legend } from './Legend';
 export { default as LineOfSight } from './LineOfSight';
 export { default as MediaMixin } from './MediaMixin';
